Task-4714:  Refactor Swagger annotations to reduce duplication
