+++ java/pom.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
@@ -20,7 +21,7 @@
     <properties>
         <!-- Stacks versions -->
         <stacks.core.commons.version>1.0.0</stacks.core.commons.version>
-        <stacks.core.api.version>1.0.0</stacks.core.api.version>
+        <stacks.core.api.version>1.0.0-SNAPSHOT</stacks.core.api.version>
         <stacks.azure.cosmos.version>1.0.0</stacks.azure.cosmos.version>
         <stacks.core.cqrs.version>1.0.0</stacks.core.cqrs.version>
         <stacks.azure-servicebus.version>1.0.0</stacks.azure-servicebus.version>
