+++ java/src/main/java/com/amido/stacks/workloads/menu/api/v1/CreateItemController.java
@@ -1,10 +1,11 @@
 package com.amido.stacks.workloads.menu.api.v1;
 
+import com.amido.stacks.core.api.annotations.CreateAPIResponses;
 import com.amido.stacks.core.api.dto.response.ResourceCreatedResponse;
 import com.amido.stacks.workloads.menu.api.v1.dto.request.CreateItemRequest;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.security.SecurityRequirement;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -13,8 +14,8 @@ import java.util.UUID;
 
 @RequestMapping(
     path = "/v1/menu/{id}/category/{categoryId}/items",
-    consumes = "application/json",
-    produces = "application/json; charset=utf-8",
+    consumes = MediaType.APPLICATION_JSON_VALUE,
+    produces = MediaType.APPLICATION_JSON_VALUE + "; charset=utf-8",
     method = RequestMethod.POST)
 public interface CreateItemController {
 
@@ -22,7 +23,6 @@ public interface CreateItemController {
   @Operation(
       tags = "Item",
       summary = "Add an item to an existing category in a menu",
-      security = @SecurityRequirement(name = "bearerAuth"),
       description = "Adds a menu item",
       operationId = "AddMenuItem")
   @CreateAPIResponses
