+++ java/src/main/java/com/amido/stacks/workloads/menu/api/v1/CreateMenuController.java
@@ -1,10 +1,11 @@
 package com.amido.stacks.workloads.menu.api.v1;
 
+import com.amido.stacks.core.api.annotations.CreateAPIResponses;
 import com.amido.stacks.core.api.dto.response.ResourceCreatedResponse;
 import com.amido.stacks.workloads.menu.api.v1.dto.request.CreateMenuRequest;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.security.SecurityRequirement;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -12,8 +13,8 @@ import javax.validation.Valid;
 
 @RequestMapping(
     path = "/v1/menu",
-    consumes = "application/json",
-    produces = "application/json; charset=utf-8",
+    consumes = MediaType.APPLICATION_JSON_VALUE,
+    produces = MediaType.APPLICATION_JSON_VALUE + "; charset=utf-8",
     method = RequestMethod.POST)
 public interface CreateMenuController {
 
@@ -21,7 +22,6 @@ public interface CreateMenuController {
   @Operation(
       tags = "Menu",
       summary = "Create a menu",
-      security = @SecurityRequirement(name = "bearerAuth"),
       description = "Adds a menu",
       operationId = "CreateMenu")
   @CreateAPIResponses
