+++ java/src/main/java/com/amido/stacks/workloads/menu/api/v1/DeleteCategoryController.java
@@ -1,19 +1,17 @@
 package com.amido.stacks.workloads.menu.api.v1;
 
+import com.amido.stacks.core.api.annotations.DeleteAPIResponses;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.security.SecurityRequirement;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.UUID;
 
-/**
- * @author ArathyKrishna
- */
 @RequestMapping(
     path = "/v1/menu/{id}/category/{categoryId}",
-    produces = "application/json; charset=utf-8",
+    produces = MediaType.APPLICATION_JSON_VALUE + "; charset=utf-8",
     method = RequestMethod.DELETE)
 public interface DeleteCategoryController {
 
@@ -21,7 +19,6 @@ public interface DeleteCategoryController {
   @Operation(
       tags = "Category",
       summary = "Removes a category and its items from menu",
-      security = @SecurityRequirement(name = "bearerAuth"),
       description = "Removes a category and its items from menu",
       operationId = "DeleteCategory")
   @DeleteAPIResponses
