+++ java/src/main/java/com/amido/stacks/workloads/menu/api/v1/DeleteMenuController.java
@@ -1,21 +1,17 @@
 package com.amido.stacks.workloads.menu.api.v1;
 
+import com.amido.stacks.core.api.annotations.DeleteAPIResponses;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.security.SecurityRequirement;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.UUID;
 
-/**
- * Controller for Deleting a menu
- *
- * @author ArathyKrishna
- */
 @RequestMapping(
     path = "/v1/menu/{id}",
-    produces = "application/json; charset=utf-8",
+    produces = MediaType.APPLICATION_JSON_VALUE + "; charset=utf-8",
     method = RequestMethod.DELETE)
 public interface DeleteMenuController {
 
@@ -23,7 +19,6 @@ public interface DeleteMenuController {
   @Operation(
       tags = "Menu",
       summary = "Removes a Menu with all it's Categories and Items",
-      security = @SecurityRequirement(name = "bearerAuth"),
       description = "Remove a menu from a restaurant",
       operationId = "DeleteMenu")
   @DeleteAPIResponses
