+++ java/src/main/java/com/amido/stacks/workloads/menu/api/v1/QueryMenuController.java
@@ -1,10 +1,15 @@
 package com.amido.stacks.workloads.menu.api.v1;
 
+import com.amido.stacks.core.api.annotations.ReadAPIResponses;
+import com.amido.stacks.core.api.annotations.SearchAPIResponses;
 import com.amido.stacks.workloads.menu.api.v1.dto.response.MenuDTO;
 import com.amido.stacks.workloads.menu.api.v1.dto.response.SearchMenuResult;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.security.SecurityRequirement;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -13,7 +18,7 @@ import java.util.UUID;
 
 @RequestMapping(
     path = "/v1/menu",
-    produces = "application/json; charset=utf-8",
+    produces = MediaType.APPLICATION_JSON_VALUE + "; charset=utf-8",
     method = RequestMethod.GET)
 public interface QueryMenuController {
 
@@ -21,7 +26,6 @@ public interface QueryMenuController {
   @Operation(
       tags = "Menu",
       summary = "Get or search a list of menus",
-      security = @SecurityRequirement(name = "bearerAuth"),
       description =
           "By passing in the appropriate options, you can search for available menus in the system")
   @SearchAPIResponses
@@ -36,9 +40,15 @@ public interface QueryMenuController {
   @Operation(
       tags = "Menu",
       summary = "Get a menu",
-      security = @SecurityRequirement(name = "bearerAuth"),
       description =
           "By passing the menu id, you can get access to available categories and items in the menu")
+  @ApiResponse(
+      responseCode = "200",
+      description = "Menu",
+      content =
+          @Content(
+              mediaType = MediaType.APPLICATION_JSON_VALUE,
+              schema = @Schema(implementation = MenuDTO.class)))
   @ReadAPIResponses
   ResponseEntity<MenuDTO> getMenu(
       @PathVariable(name = "id") UUID id,
