+++ java/src/main/java/com/amido/stacks/workloads/menu/api/v1/UpdateItemController.java
@@ -1,29 +1,27 @@
 package com.amido.stacks.workloads.menu.api.v1;
 
+import com.amido.stacks.core.api.annotations.UpdateAPIResponses;
+import com.amido.stacks.core.api.dto.response.ResourceUpdatedResponse;
 import com.amido.stacks.workloads.menu.api.v1.dto.request.UpdateItemRequest;
-import com.amido.stacks.workloads.menu.api.v1.dto.response.ResourceUpdatedResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.security.SecurityRequirement;
-import java.util.UUID;
-import javax.validation.Valid;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-/**
- * @author ArathyKrishna
- */
+import javax.validation.Valid;
+import java.util.UUID;
+
 @RequestMapping(
     path = "/v1/menu/{id}/category/{categoryId}/items/{itemId}",
-    consumes = "application/json",
-    produces = "application/json; charset=utf-8",
+    consumes = MediaType.APPLICATION_JSON_VALUE,
+    produces = MediaType.APPLICATION_JSON_VALUE + "; charset=utf-8",
     method = RequestMethod.PUT)
 public interface UpdateItemController {
   @PutMapping
   @Operation(
       tags = "Item",
       summary = "Update an item in the menu",
-      security = @SecurityRequirement(name = "bearerAuth"),
       description = "Update an item in the menu",
       operationId = "UpdateMenuItem")
   @UpdateAPIResponses
