+++ java/src/main/java/com/amido/stacks/workloads/menu/api/v1/impl/UpdateCategoryControllerImpl.java
@@ -1,16 +1,17 @@
 package com.amido.stacks.workloads.menu.api.v1.impl;
 
-import static com.amido.stacks.workloads.menu.mappers.RequestToCommandMapper.map;
-import static org.springframework.http.HttpStatus.OK;
-
+import com.amido.stacks.core.api.dto.response.ResourceUpdatedResponse;
 import com.amido.stacks.workloads.menu.api.v1.UpdateCategoryController;
 import com.amido.stacks.workloads.menu.api.v1.dto.request.UpdateCategoryRequest;
-import com.amido.stacks.workloads.menu.api.v1.dto.response.ResourceUpdatedResponse;
 import com.amido.stacks.workloads.menu.handlers.UpdateCategoryHandler;
-import java.util.UUID;
-import javax.validation.Valid;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
+
+import javax.validation.Valid;
+import java.util.UUID;
+
+import static com.amido.stacks.workloads.menu.mappers.RequestToCommandMapper.map;
+import static org.springframework.http.HttpStatus.OK;
 
 /**
  * Controller for updating category.
